Move BalanceProofSigned/UnsignedState to prevent cyclic imports
