+++ raiden/transfer/architecture.py
@@ -2,22 +2,36 @@
 from copy import deepcopy
 from dataclasses import dataclass, field
 
-from raiden.transfer.identifiers import QueueIdentifier
+from raiden.constants import EMPTY_BALANCE_HASH, UINT64_MAX, UINT256_MAX
+from raiden.transfer.identifiers import CanonicalIdentifier, QueueIdentifier
+from raiden.transfer.utils import hash_balance_data
 from raiden.utils.typing import (
+    AdditionalHash,
     Address,
     Any,
+    BalanceHash,
     BlockExpiration,
     BlockHash,
     BlockNumber,
     Callable,
+    ChainID,
     ChannelID,
     Generic,
     List,
+    Locksroot,
     MessageID,
+    Nonce,
     Optional,
+    Signature,
+    T_Address,
     T_BlockHash,
     T_BlockNumber,
     T_ChannelID,
+    T_Keccak256,
+    T_Signature,
+    T_TokenAmount,
+    TokenAmount,
+    TokenNetworkAddress,
     TransactionHash,
     Tuple,
     TypeVar,
@@ -270,3 +284,143 @@ class TransitionResult(Generic[ST]):  # pylint: disable=unsubscriptable-object
 
     def __ne__(self, other: Any) -> bool:
         return not self.__eq__(other)
+
+
+@dataclass
+class BalanceProofUnsignedState(State):
+    """ Balance proof from the local node without the signature. """
+
+    nonce: Nonce
+    transferred_amount: TokenAmount
+    locked_amount: TokenAmount
+    locksroot: Locksroot
+    canonical_identifier: CanonicalIdentifier
+    balance_hash: BalanceHash = field(default=EMPTY_BALANCE_HASH)
+
+    def __post_init__(self) -> None:
+        if not isinstance(self.nonce, int):
+            raise ValueError("nonce must be int")
+
+        if not isinstance(self.transferred_amount, T_TokenAmount):
+            raise ValueError("transferred_amount must be a token_amount instance")
+
+        if not isinstance(self.locked_amount, T_TokenAmount):
+            raise ValueError("locked_amount must be a token_amount instance")
+
+        if not isinstance(self.locksroot, T_Keccak256):
+            raise ValueError("locksroot must be a keccak256 instance")
+
+        if self.nonce <= 0:
+            raise ValueError("nonce cannot be zero or negative")
+
+        if self.nonce > UINT64_MAX:
+            raise ValueError("nonce is too large")
+
+        if self.transferred_amount < 0:
+            raise ValueError("transferred_amount cannot be negative")
+
+        if self.transferred_amount > UINT256_MAX:
+            raise ValueError("transferred_amount is too large")
+
+        if len(self.locksroot) != 32:
+            raise ValueError("locksroot must have length 32")
+
+        self.canonical_identifier.validate()
+
+        self.balance_hash = hash_balance_data(
+            transferred_amount=self.transferred_amount,
+            locked_amount=self.locked_amount,
+            locksroot=self.locksroot,
+        )
+
+    @property
+    def chain_id(self) -> ChainID:
+        return self.canonical_identifier.chain_identifier
+
+    @property
+    def token_network_identifier(self) -> TokenNetworkAddress:
+        return TokenNetworkAddress(self.canonical_identifier.token_network_address)
+
+    @property
+    def channel_identifier(self) -> ChannelID:
+        return self.canonical_identifier.channel_identifier
+
+
+@dataclass
+class BalanceProofSignedState(State):
+    """ Proof of a channel balance that can be used on-chain to resolve
+    disputes.
+    """
+
+    nonce: Nonce
+    transferred_amount: TokenAmount
+    locked_amount: TokenAmount
+    locksroot: Locksroot
+    message_hash: AdditionalHash
+    signature: Signature
+    sender: Address
+    canonical_identifier: CanonicalIdentifier
+    balance_hash: BalanceHash = field(default=EMPTY_BALANCE_HASH)
+
+    def __post_init__(self) -> None:
+        if not isinstance(self.nonce, int):
+            raise ValueError("nonce must be int")
+
+        if not isinstance(self.transferred_amount, T_TokenAmount):
+            raise ValueError("transferred_amount must be a token_amount instance")
+
+        if not isinstance(self.locked_amount, T_TokenAmount):
+            raise ValueError("locked_amount must be a token_amount instance")
+
+        if not isinstance(self.locksroot, T_Keccak256):
+            raise ValueError("locksroot must be a keccak256 instance")
+
+        if not isinstance(self.message_hash, T_Keccak256):
+            raise ValueError("message_hash must be a keccak256 instance")
+
+        if not isinstance(self.signature, T_Signature):
+            raise ValueError("signature must be a signature instance")
+
+        if not isinstance(self.sender, T_Address):
+            raise ValueError("sender must be an address instance")
+
+        if self.nonce <= 0:
+            raise ValueError("nonce cannot be zero or negative")
+
+        if self.nonce > UINT64_MAX:
+            raise ValueError("nonce is too large")
+
+        if self.transferred_amount < 0:
+            raise ValueError("transferred_amount cannot be negative")
+
+        if self.transferred_amount > UINT256_MAX:
+            raise ValueError("transferred_amount is too large")
+
+        if len(self.locksroot) != 32:
+            raise ValueError("locksroot must have length 32")
+
+        if len(self.message_hash) != 32:
+            raise ValueError("message_hash is an invalid hash")
+
+        if len(self.signature) != 65:
+            raise ValueError("signature is an invalid signature")
+
+        self.canonical_identifier.validate()
+
+        self.balance_hash = hash_balance_data(
+            transferred_amount=self.transferred_amount,
+            locked_amount=self.locked_amount,
+            locksroot=self.locksroot,
+        )
+
+    @property
+    def chain_id(self) -> ChainID:
+        return self.canonical_identifier.chain_identifier
+
+    @property
+    def token_network_identifier(self) -> TokenNetworkAddress:
+        return TokenNetworkAddress(self.canonical_identifier.token_network_address)
+
+    @property
+    def channel_identifier(self) -> ChannelID:
+        return self.canonical_identifier.channel_identifier
