+++ raiden/transfer/utils.py
@@ -2,160 +2,15 @@ import random
 from random import Random
 from typing import TYPE_CHECKING
 
-from eth_utils import to_checksum_address
 from web3 import Web3
 
 from raiden.constants import EMPTY_HASH
-from raiden.storage import sqlite
-from raiden.transfer.identifiers import CanonicalIdentifier
-from raiden.utils.serialization import serialize_bytes
-from raiden.utils.typing import (
-    Address,
-    Any,
-    BalanceHash,
-    Dict,
-    Locksroot,
-    Secret,
-    SecretHash,
-    TokenAmount,
-    Union,
-)
+from raiden.utils.typing import Any, BalanceHash, Locksroot, Secret, SecretHash, TokenAmount, Union
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import
     from raiden.transfer.mediated_transfer.state_change import ReceiveSecretReveal  # noqa: F401
     from raiden.transfer.state_change import ContractReceiveSecretReveal  # noqa: F401
-
-
-def get_state_change_with_balance_proof_by_balance_hash(
-    storage: sqlite.SQLiteStorage,
-    canonical_identifier: CanonicalIdentifier,
-    balance_hash: BalanceHash,
-    sender: Address,
-) -> sqlite.StateChangeRecord:
-    """ Returns the state change which contains the corresponding balance
-    proof.
-
-    Use this function to find a balance proof for a call to settle, which only
-    has the blinded balance proof data.
-    """
-    return storage.get_latest_state_change_by_data_field(
-        {
-            "balance_proof.canonical_identifier.chain_identifier": str(
-                canonical_identifier.chain_identifier
-            ),
-            "balance_proof.canonical_identifier.token_network_address": to_checksum_address(
-                canonical_identifier.token_network_address
-            ),
-            "balance_proof.canonical_identifier.channel_identifier": str(
-                canonical_identifier.channel_identifier
-            ),
-            "balance_proof.balance_hash": serialize_bytes(balance_hash),
-            "balance_proof.sender": to_checksum_address(sender),
-        }
-    )
-
-
-def get_state_change_with_balance_proof_by_locksroot(
-    storage: sqlite.SQLiteStorage,
-    canonical_identifier: CanonicalIdentifier,
-    locksroot: Locksroot,
-    sender: Address,
-) -> sqlite.StateChangeRecord:
-    """ Returns the state change which contains the corresponding balance
-    proof.
-
-    Use this function to find a balance proof for a call to unlock, which only
-    happens after settle, so the channel has the unblinded version of the
-    balance proof.
-    """
-    return storage.get_latest_state_change_by_data_field(
-        {
-            "balance_proof.canonical_identifier.chain_identifier": str(
-                canonical_identifier.chain_identifier
-            ),
-            "balance_proof.canonical_identifier.token_network_address": to_checksum_address(
-                canonical_identifier.token_network_address
-            ),
-            "balance_proof.canonical_identifier.channel_identifier": str(
-                canonical_identifier.channel_identifier
-            ),
-            "balance_proof.locksroot": serialize_bytes(locksroot),
-            "balance_proof.sender": to_checksum_address(sender),
-        }
-    )
-
-
-def get_event_with_balance_proof_by_balance_hash(
-    storage: sqlite.SQLiteStorage,
-    canonical_identifier: CanonicalIdentifier,
-    balance_hash: BalanceHash,
-) -> sqlite.EventRecord:
-    """ Returns the event which contains the corresponding balance
-    proof.
-
-    Use this function to find a balance proof for a call to settle, which only
-    has the blinded balance proof data.
-    """
-    filters = {
-        "canonical_identifier.chain_identifier": str(canonical_identifier.chain_identifier),
-        "canonical_identifier.token_network_address": to_checksum_address(
-            canonical_identifier.token_network_address
-        ),
-        "canonical_identifier.channel_identifier": str(canonical_identifier.channel_identifier),
-        "balance_hash": serialize_bytes(balance_hash),
-    }
-
-    event = storage.get_latest_event_by_data_field(
-        balance_proof_query_from_keys(prefix="", filters=filters)
-    )
-    if event.data:
-        return event
-
-    event = storage.get_latest_event_by_data_field(
-        balance_proof_query_from_keys(prefix="transfer.", filters=filters)
-    )
-    return event
-
-
-def get_event_with_balance_proof_by_locksroot(
-    storage: sqlite.SQLiteStorage,
-    canonical_identifier: CanonicalIdentifier,
-    locksroot: Locksroot,
-    recipient: Address,
-) -> sqlite.EventRecord:
-    """ Returns the event which contains the corresponding balance proof.
-
-    Use this function to find a balance proof for a call to unlock, which only
-    happens after settle, so the channel has the unblinded version of the
-    balance proof.
-    """
-    filters = {
-        "canonical_identifier.chain_identifier": str(canonical_identifier.chain_identifier),
-        "canonical_identifier.token_network_address": to_checksum_address(
-            canonical_identifier.token_network_address
-        ),
-        "canonical_identifier.channel_identifier": str(canonical_identifier.channel_identifier),
-        "locksroot": serialize_bytes(locksroot),
-        "recipient": to_checksum_address(recipient),
-    }
-    event = storage.get_latest_event_by_data_field(
-        balance_proof_query_from_keys(prefix="", filters=filters)
-    )
-    if event.data:
-        return event
-
-    event = storage.get_latest_event_by_data_field(
-        balance_proof_query_from_keys(prefix="transfer.", filters=filters)
-    )
-    return event
-
-
-def balance_proof_query_from_keys(prefix: str, filters: Dict[str, Any]) -> Dict[str, Any]:
-    transformed_filters = {}
-    for key, value in filters.items():
-        transformed_filters[f"{prefix}balance_proof.{key}"] = value
-    return transformed_filters
 
 
 def hash_balance_data(
